Subtract V3
