+++ src/main/java/calculatorpackage/App.java
@@ -2,8 +2,7 @@ package calculatorpackage;
 
 public class App {
 	
-	public static void main(String[] args) {
-		calculateValue();		
+	public static void main(String[] args) {	
 		
 	}
